clip_server/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
git+https://github.com/openai/CLIP.git
python-multipart
